Upload handling fix

+++ src/app/api/uploadthing/core.ts
@@ -12,26 +12,20 @@ export const fileRouter = {
   })
     .middleware(async () => {
       const { user } = await validateRequest();
-
       if (!user) throw new UploadThingError("Unauthorized");
-
       return { user };
     })
     .onUploadComplete(async ({ metadata, file }) => {
       const oldAvatarUrl = metadata.user.avatarUrl;
 
       if (oldAvatarUrl) {
-        const key = oldAvatarUrl.split(
-          `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
-        )[1];
-
-        await new UTApi().deleteFiles(key);
+        const key = oldAvatarUrl.split("/f/").pop();
+        if (key) {
+          await new UTApi().deleteFiles(key);
+        }
       }
 
-      const newAvatarUrl = file.url.replace(
-        "/f/",
-        `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
-      );
+      const newAvatarUrl = file.url;
 
       await Promise.all([
         prisma.user.update({
@@ -48,24 +42,20 @@ export const fileRouter = {
 
       return { avatarUrl: newAvatarUrl };
     }),
+
   attachment: f({
     image: { maxFileSize: "4MB", maxFileCount: 5 },
     video: { maxFileSize: "16MB", maxFileCount: 5 },
   })
     .middleware(async () => {
       const { user } = await validateRequest();
-
       if (!user) throw new UploadThingError("Unauthorized");
-
       return {};
     })
     .onUploadComplete(async ({ file }) => {
       const media = await prisma.media.create({
         data: {
-          url: file.url.replace(
-            "/f/",
-            `/a/${process.env.NEXT_PUBLIC_UPLOADTHING_APP_ID}/`,
-          ),
+          url: file.url,
           type: file.type.startsWith("image") ? "IMAGE" : "VIDEO",
         },
       });
